Move shard key mapping type definition into separate module
